% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_functions.R
\docType{package}
\name{siteVerification_googleAuthR}
\alias{siteVerification_googleAuthR}
\alias{siteVerification_googleAuthR-package}
\title{Google Site Verification API
Verifies ownership of websites or domains with Google.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:47:15
filename: /Users/mark/dev/R/autoGoogleAPI/googlesiteVerificationv1.auto/R/siteVerification_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/siteverification
\item https://www.googleapis.com/auth/siteverification.verify_only
}
}

