% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceGettokenRequest}
\alias{SiteVerificationWebResourceGettokenRequest}
\title{Google Site Verification API Objects 
Verifies ownership of websites or domains with Google.}
\usage{

  SiteVerificationWebResourceGettokenRequest(SiteVerificationWebResourceGettokenRequest.site = NULL,
  site = NULL, verificationMethod = NULL)
}
\arguments{
\item{SiteVerificationWebResourceGettokenRequest.site}{The \link{SiteVerificationWebResourceGettokenRequest.site} object or list of objects}

\item{site}{The site for which a verification token will be generated}

\item{verificationMethod}{The verification method that will be used to verify this site}
}
\value{
SiteVerificationWebResourceGettokenRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:13:57
filename: /Users/mark/dev/R/autoGoogleAPI/googlesiteVerificationv1.auto/R/siteVerification_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SiteVerificationWebResourceGettokenRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SiteVerificationWebResourceGettokenRequest functions: \code{\link{SiteVerificationWebResourceGettokenRequest.site}},
  \code{\link{webResource.getToken}}
}
