% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceGettokenRequest.site}
\alias{SiteVerificationWebResourceGettokenRequest.site}
\title{SiteVerificationWebResourceGettokenRequest.site Object}
\usage{
SiteVerificationWebResourceGettokenRequest.site(identifier = NULL,
  type = NULL)
}
\arguments{
\item{identifier}{The site identifier}

\item{type}{The type of resource to be verified}
}
\value{
SiteVerificationWebResourceGettokenRequest.site object
}
\description{
SiteVerificationWebResourceGettokenRequest.site Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The site for which a verification token will be generated.
}
\seealso{
Other SiteVerificationWebResourceGettokenRequest functions: \code{\link{SiteVerificationWebResourceGettokenRequest}},
  \code{\link{webResource.getToken}}
}
