% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceGettokenResponse}
\alias{SiteVerificationWebResourceGettokenResponse}
\title{SiteVerificationWebResourceGettokenResponse Object}
\usage{
SiteVerificationWebResourceGettokenResponse(method = NULL, token = NULL)
}
\arguments{
\item{method}{The verification method to use in conjunction with this token}

\item{token}{The verification token}
}
\value{
SiteVerificationWebResourceGettokenResponse object
}
\description{
SiteVerificationWebResourceGettokenResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
