% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceListResponse}
\alias{SiteVerificationWebResourceListResponse}
\title{SiteVerificationWebResourceListResponse Object}
\usage{
SiteVerificationWebResourceListResponse(items = NULL)
}
\arguments{
\item{items}{The list of sites that are owned by the authenticated user}
}
\value{
SiteVerificationWebResourceListResponse object
}
\description{
SiteVerificationWebResourceListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
