% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceResource}
\alias{SiteVerificationWebResourceResource}
\title{SiteVerificationWebResourceResource Object}
\usage{

  SiteVerificationWebResourceResource(SiteVerificationWebResourceResource.site = NULL,
  id = NULL, owners = NULL, site = NULL)
}
\arguments{
\item{SiteVerificationWebResourceResource.site}{The \link{SiteVerificationWebResourceResource.site} object or list of objects}

\item{id}{The string used to identify this site}

\item{owners}{The email addresses of all verified owners}

\item{site}{The address and type of a site that is verified or will be verified}
}
\value{
SiteVerificationWebResourceResource object
}
\description{
SiteVerificationWebResourceResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other SiteVerificationWebResourceResource functions: \code{\link{SiteVerificationWebResourceResource.site}},
  \code{\link{webResource.insert}},
  \code{\link{webResource.patch}},
  \code{\link{webResource.update}}
}
