% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_objects.R
\name{SiteVerificationWebResourceResource.site}
\alias{SiteVerificationWebResourceResource.site}
\title{SiteVerificationWebResourceResource.site Object}
\usage{
SiteVerificationWebResourceResource.site(identifier = NULL, type = NULL)
}
\arguments{
\item{identifier}{The site identifier}

\item{type}{The site type}
}
\value{
SiteVerificationWebResourceResource.site object
}
\description{
SiteVerificationWebResourceResource.site Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The address and type of a site that is verified or will be verified.
}
\seealso{
Other SiteVerificationWebResourceResource functions: \code{\link{SiteVerificationWebResourceResource}},
  \code{\link{webResource.insert}},
  \code{\link{webResource.patch}},
  \code{\link{webResource.update}}
}

