% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_functions.R
\name{webResource.getToken}
\alias{webResource.getToken}
\title{Get a verification token for placing on a website or domain.}
\usage{
webResource.getToken(SiteVerificationWebResourceGettokenRequest)
}
\arguments{
\item{SiteVerificationWebResourceGettokenRequest}{The \link{SiteVerificationWebResourceGettokenRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/siteverification
\item https://www.googleapis.com/auth/siteverification.verify_only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/siteverification, https://www.googleapis.com/auth/siteverification.verify_only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/site-verification/}{Google Documentation}

Other SiteVerificationWebResourceGettokenRequest functions: \code{\link{SiteVerificationWebResourceGettokenRequest.site}},
  \code{\link{SiteVerificationWebResourceGettokenRequest}}
}

