% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_functions.R
\name{webResource.insert}
\alias{webResource.insert}
\title{Attempt verification of a website or domain.}
\usage{
webResource.insert(SiteVerificationWebResourceResource, verificationMethod)
}
\arguments{
\item{SiteVerificationWebResourceResource}{The \link{SiteVerificationWebResourceResource} object to pass to this method}

\item{verificationMethod}{The method to use for verifying a site or domain}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/siteverification
\item https://www.googleapis.com/auth/siteverification.verify_only
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/siteverification, https://www.googleapis.com/auth/siteverification.verify_only)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/site-verification/}{Google Documentation}

Other SiteVerificationWebResourceResource functions: \code{\link{SiteVerificationWebResourceResource.site}},
  \code{\link{SiteVerificationWebResourceResource}},
  \code{\link{webResource.patch}},
  \code{\link{webResource.update}}
}
