% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_functions.R
\name{webResource.list}
\alias{webResource.list}
\title{Get the list of your verified websites and domains.}
\usage{
webResource.list()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/siteverification
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/siteverification)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/site-verification/}{Google Documentation}
}

