% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteVerification_functions.R
\name{webResource.update}
\alias{webResource.update}
\title{Modify the list of owners for your website or domain.}
\usage{
webResource.update(SiteVerificationWebResourceResource, id)
}
\arguments{
\item{SiteVerificationWebResourceResource}{The \link{SiteVerificationWebResourceResource} object to pass to this method}

\item{id}{The id of a verified site or domain}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/siteverification
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/siteverification)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/site-verification/}{Google Documentation}

Other SiteVerificationWebResourceResource functions: \code{\link{SiteVerificationWebResourceResource.site}},
  \code{\link{SiteVerificationWebResourceResource}},
  \code{\link{webResource.insert}},
  \code{\link{webResource.patch}}
}

